requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.14.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
